requirements.txt
python-dotenv>=1.2.1
langfuse>=3.8.1
nltk>=3.9.2
numpy>=2.3.4
torch>=2.9.0
openai>=2.6.1
transformers>=4.57.1
langchain-core>=1.0.1
langchain-google-genai>=3.0.0
langchain-openai>=1.0.1
pydantic>=2.12.3